- [sqlite driver] mitigated a memory leak.
When a  lot of queries  are performed the  memory usage of  the system
keeps growing.

Seems that this  was related to the  fact that the driver  did not not
called "finalize-statement" before.  I have added a few  of this forms
and at least the problem has been reduced a lot.

Tested with a million of "SELECTs" on a file database.