- modified the value of 'ping' expression for sqlite driver.
I think  would be  useful to  tell apart  if a  sqlite database  is in
memory or  an actual file  and, if the latter  is true, give  the path
file to the database the connection is linked to.

I think this changes should be harmless for users that rely on the old
behaviour of the that function.