Created ability for separate start-transaction
In oder to move the start-transaction, commit, rollback to a higher
level had to create a separate start-transaction function which sets
the *transaction-state* and then calls the begin-transaction at the
driver level.  This means there also has to be an end-transaction
which clears the *transaction-state* because the macros rely on the
dynamic variable structure of *transaction-state*.  Also fixed the
release-savepoint which was sending the wrong message.